third_party/blink/web_tests/external/wpt/mathml/presentation-markup/mrow/spacing-expected.txt

This is a testharness.js-based test.
PASS operator spacing inside Mrow
PASS operator spacing inside Sqrt
PASS operator spacing inside Style
PASS operator spacing inside Error
PASS operator spacing inside Phantom
PASS operator spacing inside Math
FAIL operator spacing inside Menclose assert_greater_than_equal: expected a number greater than or equal to 50 but got -8
PASS operator spacing inside Mpadded
FAIL operator spacing inside Unknown assert_greater_than_equal: expected a number greater than or equal to 50 but got -8
Harness: the test ran to completion.
